+++ src/main/java/com/github/sgoeschl/freemarker/cli/tools/properties/PropertiesTool.java
@@ -32,7 +32,7 @@ public class PropertiesTool {
     }
 
     public Properties parse(Document document) {
-        try (InputStream is = document.getInputStream()) {
+        try (InputStream is = document.getUnsafeInputStream()) {
             final Properties properties = new Properties();
             properties.load(is);
             return properties;