+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file. We try to a
 ## 2.0.0-BETA-5-SNAPSHOT
 
 ### Added
+- `Document#getUnsafeInputStream` to get a input stream which needs to be closed by the caller
+- `CSVTool` allows parsing plain string CSV
 - Tools are loaded from `freemarker-cli.properties` and freely configurable
 - Add RAT & PMD report to Maven site generation
 - Add `-times` command line option to execute the transformation N times